fixed TrecWebParser
The TrecWebParser was falling down due to malformed document contents
(for example, the final doc in WTX001/B01 ends in the middle of an open
anchor tag). While the lxml parser was able to handle the malformed html
better than the default html.parser, it was somehow missing several docs
while parsing the full trecweb formatted file. This new approach uses
simple string manipulations to split apart the individual <DOC>
elements, then stores only the contents of the files, without the
<DOCHDR> or any other elements.